requirements.txt
alembic
fastapi
sqlalchemy
asyncpg
uvicorn
passlib
python-jose
pydantic-settings
python-multipart
psycopg2
